Adding support to specify which disks belong to which ring
Updating README for the tool

Update ringdisk logic to support dict and list

We want to be able to qualify which rings get placed on which disks.
This change enables that but also supports backwards compatibility
with all rings being on all data disks.

Update README.md

Adding start swift services after rings are copied